Switch to Facebook login

+++ AutoTag-API/src/project.ts
@@ -6,7 +6,8 @@ export const router: Router = Router();
 router.get('/project', async function (req: Request, res: Response, next: NextFunction) {
   try {
     const repository = await getProjectRepository();
-    const allProjects = await repository.find();
+    const allProjects = await repository.find({ownerUid: req.user.uid});
+    console.log(req.user.email + ': Get all projects');
     res.send(allProjects);
   }
   catch (err) {
@@ -17,7 +18,8 @@ router.get('/project', async function (req: Request, res: Response, next: NextFu
 router.get('/project/:id', async function (req: Request, res: Response, next: NextFunction) {
   try {
     const repository = await getProjectRepository();
-    const project = await repository.find({id: Number(req.params.id)});
+    const project = await repository.find({id: Number(req.params.id), ownerUid: req.user.uid});
+    console.log(req.user.email + ': Get project ' + req.params.id);
     res.send(project);
   }
   catch (err) {
@@ -29,12 +31,14 @@ router.post('/project', async function (req: Request, res: Response, next: NextF
   try {
     const repository = await getProjectRepository();
     const project = new Project();
+    project.ownerUid = req.user.uid;
     project.name = req.body.name;
     project.description = req.body.description;
     project.createDate = new Date();
     project.editDate = new Date();
 
     const result = await repository.save(project);
+    console.log(req.user.email + ': Create project ' + project.id);
     res.send(result);
   }
   catch (err) {
@@ -45,12 +49,13 @@ router.post('/project', async function (req: Request, res: Response, next: NextF
 router.post('/project/:id', async function (req: Request, res: Response, next: NextFunction) {
   try {
     const repository = await getProjectRepository();
-    const project = await repository.findOne({id: Number(req.params.id)});
+    const project = await repository.findOne({id: Number(req.params.id), ownerUid: req.user.uid});
     project.name = req.body.name;
     project.description = req.body.description;
     project.editDate = new Date();
 
     const result = await repository.save(project);
+    console.log(req.user.email + ': Update project ' + req.params.id);
     res.send(result);
   }
   catch (err) {
@@ -61,7 +66,8 @@ router.post('/project/:id', async function (req: Request, res: Response, next: N
 router.delete('/project/:id', async function (req: Request, res: Response, next: NextFunction) {
   try {
     const repository = await getProjectRepository();
-    await repository.delete({id: Number(req.params.id)});
+    await repository.delete({id: Number(req.params.id), ownerUid: req.user.uid});
+    console.log(req.user.email + ': Delete project ' + req.params.id);
     res.send('OK');
   }
   catch (err) {
